commit code update phân quyền

+++ resources/views/admin/pages/actors/index.blade.php
@@ -93,11 +93,13 @@
                             <div class="card-header">
                                 <h4 class="card-title">Danh sách diễn viên</h4>
                                 <div class="compose-btn">
-                                    <a href="{{ route('admin.actors.create') }}">
-                                        <button class="btn btn-secondary btn-sm light">
-                                            + Thêm mới
-                                        </button>
-                                    </a>
+                                    @can('create', \App\Models\Actor::class)
+                                        <a href="{{ route('admin.actors.create') }}">
+                                            <button class="btn btn-secondary btn-sm light">
+                                                + Thêm mới
+                                            </button>
+                                        </a>
+                                    @endcan
                                 </div>
                             </div>
                             <div class="card-body">
@@ -106,37 +108,43 @@
                                         <table class="table table-responsive-md" id="data-table">
                                             <thead>
                                                 <tr>
-                                                    <th>
-                                                        <div class="box-delete-item">
-                                                            <input type="checkbox" id="item-all-checked">
-                                                            <button id="btn-delete-all"
-                                                                data-url="{{ route('admin.actors.deleteItemMultipleChecked') }}"
-                                                                class="btn btn-sm btn-danger btn-delete-multiple_item">
-                                                                <svg width="15" height="15"
-                                                                    xmlns="http://www.w3.org/2000/svg"
-                                                                    viewBox="0 0 448 512">
-                                                                    <!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.-->
-                                                                    <path fill="white"
-                                                                        d="M135.2 17.7C140.6 6.8 151.7 0 163.8 0L284.2 0c12.1 0 23.2 6.8 28.6 17.7L320 32l96 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 96C14.3 96 0 81.7 0 64S14.3 32 32 32l96 0 7.2-14.3zM32 128l384 0 0 320c0 35.3-28.7 64-64 64L96 512c-35.3 0-64-28.7-64-64l0-320zm96 64c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16z" />
-                                                                </svg>
-                                                            </button>
-                                                        </div>
-                                                    </th>
+                                                    @can('deleteMultiple', \App\Models\Actor::class)
+                                                        <th>
+                                                            <div class="box-delete-item">
+                                                                <input type="checkbox" id="item-all-checked">
+                                                                <button id="btn-delete-all"
+                                                                    data-url="{{ route('admin.actors.deleteItemMultipleChecked') }}"
+                                                                    class="btn btn-sm btn-danger btn-delete-multiple_item">
+                                                                    <svg width="15" height="15"
+                                                                        xmlns="http://www.w3.org/2000/svg"
+                                                                        viewBox="0 0 448 512">
+                                                                        <!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.-->
+                                                                        <path fill="white"
+                                                                            d="M135.2 17.7C140.6 6.8 151.7 0 163.8 0L284.2 0c12.1 0 23.2 6.8 28.6 17.7L320 32l96 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 96C14.3 96 0 81.7 0 64S14.3 32 32 32l96 0 7.2-14.3zM32 128l384 0 0 320c0 35.3-28.7 64-64 64L96 512c-35.3 0-64-28.7-64-64l0-320zm96 64c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16zm96 0c-8.8 0-16 7.2-16 16l0 224c0 8.8 7.2 16 16 16s16-7.2 16-16l0-224c0-8.8-7.2-16-16-16z" />
+                                                                    </svg>
+                                                                </button>
+                                                            </div>
+                                                        </th>
+                                                    @endcan
                                                     <th style="width:80px;">#</th>
                                                     <th>Tên</th>
                                                     <th>Hình ảnh</th>
                                                     <th>Ngày sinh</th>
                                                     <th>Quốc tịch</th>
-                                                    <th>Hành động</th>
+                                                    @if (Auth()->user()->can('update', \App\Models\Actor::class) || Auth()->user()->can('delete', \App\Models\Actor::class))
+                                                        <th>Hành động</th>
+                                                    @endif
                                                 </tr>
                                             </thead>
                                             <tbody>
                                                 @foreach ($data as $key => $actor)
                                                     <tr>
-                                                        <td>
-                                                            <input type="checkbox" data-id="{{ $actor->id }}"
-                                                                class="item-checked">
-                                                        </td>
+                                                        @can('deleteMultiple', \App\Models\Actor::class)
+                                                            <td>
+                                                                <input type="checkbox" data-id="{{ $actor->id }}"
+                                                                    class="item-checked">
+                                                            </td>
+                                                        @endcan
                                                         <td>
                                                             <strong class="text-black">{{ ($data->currentPage() - 1) * $data->perPage() + $key + 1 }}</strong>
                                                         </td>
@@ -166,26 +174,32 @@
                                                                 {{ $actor->nationality }}
                                                             </b>
                                                         </td>
-                                                        <td>
-                                                            <div
-                                                                style="padding-right: 20px; display: flex; justify-content: end">
-                                                                <a href="{{ route('admin.actors.edit', $actor->id) }}"
-                                                                    class="btn btn-primary shadow btn-xs sharp me-1">
-                                                                    <i class="fa fa-pencil"></i>
-                                                                </a>
-                                                                <form
-                                                                    action="{{ route('admin.actors.delete', $actor->id) }}"
-                                                                    class="formDelete" method="post">
-                                                                    @csrf
-                                                                    @method('DELETE')
-                                                                    <button
-                                                                        class="btn btn-danger shadow btn-xs sharp me-1 call-ajax btn-remove btnDelete"
-                                                                        data-type="DELETE" href="">
-                                                                        <i class="fa fa-trash"></i>
-                                                                    </button>
-                                                                </form>
-                                                            </div>
-                                                        </td>
+                                                        @if (Auth()->user()->can('update', \App\Models\Actor::class) || Auth()->user()->can('delete', \App\Models\Actor::class))
+                                                            <td>
+                                                                <div
+                                                                    style="padding-right: 20px; display: flex; justify-content: end">
+                                                                    @can('update', \App\Models\Actor::class)
+                                                                        <a href="{{ route('admin.actors.edit', $actor->id) }}"
+                                                                            class="btn btn-primary shadow btn-xs sharp me-1">
+                                                                            <i class="fa fa-pencil"></i>
+                                                                        </a>
+                                                                    @endcan
+                                                                    @can('delete', \App\Models\Actor::class)
+                                                                        <form
+                                                                            action="{{ route('admin.actors.delete', $actor->id) }}"
+                                                                            class="formDelete" method="post">
+                                                                            @csrf
+                                                                            @method('DELETE')
+                                                                            <button
+                                                                                class="btn btn-danger shadow btn-xs sharp me-1 call-ajax btn-remove btnDelete"
+                                                                                data-type="DELETE" href="">
+                                                                                <i class="fa fa-trash"></i>
+                                                                            </button>
+                                                                        </form>
+                                                                    @endcan
+                                                                </div>
+                                                            </td>
+                                                        @endif
                                                     </tr>
                                                 @endforeach
                                             </tbody>
